# swagger.yaml
openapi: 3.0.0
info:
  title: CarsForms Database API Read Only
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: http://172.16.202.63:5400
    description: Optional server description, e.g. Main (production) server
  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing
paths:
  /api/cars_all:
    get:
      summary: Returns a list of cars all.
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string

  /api/cars/{section}/{car}/{year}/{month}:
    get:
      summary: Returns for DashBoard Summary of Cars Form.
      parameters:
        - in: path
          name: section
          schema:
            type: string
          required: true
          description: The section
        - in: path
          name: car
          schema:
            type: string
          required: true
          description: The car
        - in: path
          name: year
          schema:
            type: string
          required: true
          description: The year
        - in: path
          name: month
          schema:
            type: string
          required: true
          description: The month 
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string 

  /api/license/{section}:
    get:
      summary: Returns a list of license for section.
      parameters:
        - in: path
          name: section
          schema:
            type: string
          required: true
          description: The section
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string


  /api/license/status/{status}:
    get:
      summary: Returns a station and range for day.
      parameters:
        - in: path
          name: status
          schema:
            type: string
          required: true
          description: The status
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string

  /api/license/check/insert/{license}:
    get:
      summary: Returns a station and range for day.
      parameters:
        - in: path
          name: license
          schema:
            type: string
          required: true
          description: The license
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string

  /api/miles/check/update/{license}: 
    get:
      summary: Returns a station and range for day.
      parameters:
        - in: path
          name: license
          schema:
            type: string
          required: true
          description: The license
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string

  /api/departments/{department}/{divition}:  
    get:
      summary: Returns a station and range for day.
      parameters:
        - in: path
          name: department
          schema:
            type: string
          required: true
          description: The department
        - in: path
          name: divition
          schema:
            type: string
          required: true
          description: The divition
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string

  /api/departments/{department}:
    get:
      summary: Returns a station and range for day.
      parameters:
        - in: path
          name: department
          schema:
            type: string
          required: true
          description: The department
      description: Optional extended description in CommonMark or HTML.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string

 
  custom-CORS-response:
    responses:
      default:
        description: >
          If you specify an OPTIONS operation in your Swagger spec, then
          Swagger-Server will honor it. You can override any of
          Swagger-Server's default values for CORS headers, as shown below.
          Or you can disable Swagger-Server's automatic CORS functionality
          entirely with `server.settings.enableCORS = false;`.
        headers:
          Access-Control-Allow-Origin:
            type: string
            default: "*"
            description: >
              Setting this header to `*` allows all origins.
              This is handy for public REST APIs that don't require
              authentication. But, according to the HTTP spec, browsers
              *WILL NOT* send cookies if this header is `*`, regardless
              of what you set `Access-Control-Allow-Credentials` to.
          Access-Control-Allow-Credentials:
            type: boolean
            default: false
            description: >
              Setting this header to `false` means that your API does not
              use authentication cookies.
          Access-Control-Allow-Headers:
            type: array
            collectionFormat: csv
            default: Content-Type, X-Powered-By
            description: >
              This shows how you can explicitly specify which HTTP headers
              your API allows.
          Access-Control-Allow-Methods:
            type: array
            collectionFormat: csv
            default: GET, HEAD
            description: >
              This shows how you can explicitly specify which HTTP methods
              your API allows.
          Access-Control-Max-Age:
            type: integer
            default: 86400 # one day
            description: >
              This allows client browsers to cache the CORS response for
              one day (86400 seconds).